buggy-app/requirements.txt
flask==3.1.1
gunicorn==23.0.0
azure-monitor-opentelemetry==1.6.4
